+++ cli/docs/changelog.md
@@ -3,6 +3,8 @@
 * Upgraded Ubuntu machine image in Circle CI configuration to version `ubuntu-2004:202201-02`
 * Upgraded goreleaser to v1.9.1 in Circle CI config file
 * Updated `build cli script` and `cli launch script` to support the goamd64 env var
+### Features
+* The gateway will reconnect if the engine pod it's connected to is restarted
 
 # 0.21.0
 ### Breaking Changes